Added methods to phononH0 class to retrieve polar information. Added method to handle indeces of the phonon eigenvector in matrix form. Split the implementation of the PhononH0 class in two files, for faster compilation.
