+++ CMakeLists.txt
@@ -20,7 +20,7 @@ endif()
 
 ############### SOURCE ###############
 
-SET (SOURCE_FILES src/bands/bandstructure.cpp src/apps/app.cpp src/apps/phonon_transport_app.cpp src/apps/polarization_app.cpp src/constants/periodic_table.cpp src/utilities.cpp src/exceptions/exceptions.cpp src/io.cpp src/particle.cpp src/context.cpp src/crystal.cpp src/harmonic/harmonic.cpp src/harmonic/phonon_h0.cpp src/points/points.cpp src/bands/active_bandstructure.cpp src/harmonic/electron_h0_fourier.cpp src/harmonic/electron_h0_wannier.cpp src/bands/window.cpp src/bte/vector_bte.cpp src/bte/basevector_bte.cpp src/parser/qe_input_parser.cpp src/observable/observable.cpp src/bte/scattering.cpp src/bte/ph_scattering.cpp src/bte/drift.cpp src/delta_function.cpp src/statistics_sweep.cpp src/parser/ifc3_parser.cpp src/interaction/interaction_3ph.cpp src/apps/dos_app.cpp src/apps/bands_app.cpp src/observable/phonon_thermal_cond.cpp src/observable/phonon_viscosity.cpp src/observable/specific_heat.cpp src/mpi/mpiHelper.cpp src/mpi/mpiController.cpp src/algebra/PMatrix.cpp src/algebra/Matrix.cpp src/observable/wigner_phonon_thermal_cond.cpp src/observable/onsager.cpp src/bte/helper_3rd_state.cpp src/points/active_points.cpp src/points/full_points.cpp src/points/path_points.cpp src/points/irreducible_points.cpp src/interaction/interaction_elph.cpp src/apps/electron_wannier_transport_app.cpp src/bte/el_scattering.cpp src/bte/helper_el_scattering.cpp)
+SET (SOURCE_FILES src/bands/bandstructure.cpp src/apps/app.cpp src/apps/phonon_transport_app.cpp src/apps/polarization_app.cpp src/constants/periodic_table.cpp src/utilities.cpp src/exceptions/exceptions.cpp src/io.cpp src/particle.cpp src/context.cpp src/crystal.cpp src/harmonic/harmonic.cpp src/harmonic/phonon_h0.cpp src/harmonic/phonon_h0_sum_rule.cpp src/points/points.cpp src/bands/active_bandstructure.cpp src/harmonic/electron_h0_fourier.cpp src/harmonic/electron_h0_wannier.cpp src/bands/window.cpp src/bte/vector_bte.cpp src/bte/basevector_bte.cpp src/parser/qe_input_parser.cpp src/observable/observable.cpp src/bte/scattering.cpp src/bte/ph_scattering.cpp src/bte/drift.cpp src/delta_function.cpp src/statistics_sweep.cpp src/parser/ifc3_parser.cpp src/interaction/interaction_3ph.cpp src/apps/dos_app.cpp src/apps/bands_app.cpp src/observable/phonon_thermal_cond.cpp src/observable/phonon_viscosity.cpp src/observable/specific_heat.cpp src/mpi/mpiHelper.cpp src/mpi/mpiController.cpp src/algebra/PMatrix.cpp src/algebra/Matrix.cpp src/observable/wigner_phonon_thermal_cond.cpp src/observable/onsager.cpp src/bte/helper_3rd_state.cpp src/points/active_points.cpp src/points/full_points.cpp src/points/path_points.cpp src/points/irreducible_points.cpp src/interaction/interaction_elph.cpp src/apps/electron_wannier_transport_app.cpp src/bte/el_scattering.cpp src/bte/helper_el_scattering.cpp)
 FILE(GLOB TEST_SOURCES test/*.cpp test/interaction3ph/*.cpp)
 
 include_directories(include)
@@ -105,7 +105,7 @@ else()
 
 endif()
 
-option(OMP_AVAIL "Build with OMP" OFF)
+option(OMP_AVAIL "Build with OMP" ON)
 if(OMP_AVAIL)
         find_package(OpenMP)
         target_link_libraries(phoebe OpenMP::OpenMP_CXX)
